requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4

grpcio==1.67.1
grpcio-tools==1.67.1
protobuf==5.28.3

langchain-openai==0.2.14
langchain-community==0.3.14
langchain-text-splitters==0.3.5
langchain-core==0.3.29
chromadb==0.5.23
pypdf==5.1.0